For Navigation: Just added comment to remind me that RoutinesScreen can route to self for recomposition if needed
For Theme: Added code for drawing the wave images at that bottom of screen so it can be reused on multiple screens.

For AllWorkouts and LocalStorage: created functions to delete workout files from local storage as well as remove workouts from the allWorkouts ArrayList (NEEDS REVIEW AS THIS COULD INTRODUCE BUGS IF LOGIC IS TIED TO THE ARRAYLIST INDEX FOR A WORKOUT).

For RoutineScreen: Added ability to delete workouts and have recomposition occur without needed to route back to itself. General code cleanup.